win32: merge previous work by @seisman into io.win
which was renamed from io.datamark in the meantime
