+++ misc/docs/source/packages/index.rst
@@ -105,7 +105,6 @@ categories.*
     obspy.io.sh
     obspy.io.wav
     obspy.io.win
-    obspy.io.win32
     obspy.io.y
 
 .. rubric:: Event Data Import/Export Plug-ins
