--- misc/docs/source/packages/obspy.io.win32.rst
@@ -1,14 +1,0 @@
-.. currentmodule:: obspy.io.win32
-.. automodule:: obspy.io.win32
-
-    .. comment to end block
-
-    Modules
-    -------
-    .. autosummary::
-       :toctree: autogen
-       :nosignatures:
-
-       core
-
-    .. comment to end block
