+++ obspy/io/win/__init__.py
@@ -2,12 +2,12 @@
 """
 obspy.io.win - WIN read support for ObsPy
 =========================================
-This module provides read support for WIN waveform data. This format is
-written by different dataloggers, including Hakusan LS-7000XT Datamark
+This module provides read support for WIN and WIN32 waveform data. This format
+is written by different dataloggers, including Hakusan LS-7000XT Datamark
 dataloggers. There are two subformats, A0 and A1. To our knowledge, A0 is the
-only one supported with the current code. A0 conforms to the
-data format of the WIN system developed by Earthquake Research Institute
-(ERI), the University of Tokyo.
+only one supported with the current code. A0 conforms to the data format of the
+WIN system developed by Earthquake Research Institute (ERI), the University of
+Tokyo.
 
 :copyright:
     The ObsPy Development Team (devs@obspy.org), Thomas Lecocq, Adolfo Inza &
