+++ obspy/io/win/docs/channel_table.txt
