--- obspy/io/win32/README.txt
@@ -1,26 +1,0 @@
-package obspy.io.win32
-======================
-
-Copyright
----------
-GNU Lesser General Public License, Version 3 (LGPLv3)
-
-Copyright (c) 2017 by:
-    * Dongdong Tian
-
-
-Overview
---------
-WIN32 read support for ObsPy.
-
-This module provides read support for WIN32 data format as defined by NIED
-(https://hinetwww11.bosai.go.jp/auth/manual/dlDialogue.php?r=win32format&LANG=en).
-
-ObsPy is an open-source project dedicated to provide a Python framework for
-processing seismological data. It provides parsers for common file formats and
-seismological signal processing routines which allow the manipulation of
-seismological time series (see Beyreuther et al. 2010, Megies et al. 2011).
-The goal of the ObsPy project is to facilitate rapid application development
-for seismology.
-
-For more information visit https://www.obspy.org.
