--- obspy/io/win32/__init__.py
@@ -1,21 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-obspy.io.win32 - WIN32 read support for ObsPy
-=============================================
-This module provides read support for WIN32 waveform data.
-
-:copyright:
-    The ObsPy Development Team (devs@obspy.org) & Dongdong Tian
-:license:
-    GNU Lesser General Public License, Version 3
-    (https://www.gnu.org/copyleft/lesser.html)
-
-"""
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from future.builtins import *  # NOQA
-
-
-if __name__ == '__main__':
-    import doctest
-    doctest.testmod(exclude_empty=True)
