+++ setup.py
@@ -149,6 +149,7 @@ ENTRY_POINTS = {
         'PICKLE = obspy.core.stream',
         'CSS = obspy.io.css.core',
         'WIN = obspy.io.win.core',
+        'WIN32 = obspy.io.win.core',
         'KINEMETRICS_EVT = obspy.io.kinemetrics.core',
         'GSE1 = obspy.io.gse2.core',
         'GSE2 = obspy.io.gse2.core',
@@ -165,7 +166,6 @@ ENTRY_POINTS = {
         'Q = obspy.io.sh.core',
         'SH_ASC = obspy.io.sh.core',
         'WAV = obspy.io.wav.core',
-        'WIN32 = obspy.io.win32.core',
         'AH = obspy.io.ah.core',
         'KNET = obspy.io.nied.knet',
         'GCF = obspy.io.gcf.core',
@@ -199,6 +199,10 @@ ENTRY_POINTS = {
     'obspy.plugin.waveform.WIN': [
         'isFormat = obspy.io.win.core:_is_win',
         'readFormat = obspy.io.win.core:_read_win',
+        ],
+    'obspy.plugin.waveform.WIN32': [
+        'isFormat = obspy.io.win.core:_is_win32',
+        'readFormat = obspy.io.win.core:_read_win32',
         ],
     'obspy.plugin.waveform.KINEMETRICS_EVT': [
         'isFormat = obspy.io.kinemetrics.core:is_evt',
@@ -264,10 +268,6 @@ ENTRY_POINTS = {
         'isFormat = obspy.io.wav.core:_is_wav',
         'readFormat = obspy.io.wav.core:_read_wav',
         'writeFormat = obspy.io.wav.core:_write_wav',
-        ],
-    'obspy.plugin.waveform.WIN32': [
-        'isFormat = obspy.io.win32.core:_is_win32',
-        'readFormat = obspy.io.win32.core:_read_win32',
         ],
     'obspy.plugin.waveform.Y': [
         'isFormat = obspy.io.y.core:_is_y',
